+++ packages/gamut-labs/src/landingPage/CTA.tsx
@@ -1,27 +1,25 @@
 import React from 'react';
 import styled from '@emotion/styled';
 
-import { ButtonDeprecated } from '@codecademy/gamut';
+import { CTAButton } from '@codecademy/gamut';
 
 const CTA = styled.div`
   margin: 2rem 0 0;
 `;
 
-export type LandingPageSectionCTAProps = {
+export type PageSectionCTAProps = {
   href: string;
   className?: string;
   testId?: string;
 };
 
-export const LandingPageSectionCTA: React.FC<LandingPageSectionCTAProps> = ({
+export const PageSectionCTA: React.FC<PageSectionCTAProps> = ({
   href,
   className,
   testId,
   children,
 }) => (
   <CTA className={className} data-testid={testId}>
-    <ButtonDeprecated theme="hyper" href={href}>
-      {children}
-    </ButtonDeprecated>
+    <CTAButton href={href}>{children}</CTAButton>
   </CTA>
 );