+++ packages/markdown-overrides/CHANGELOG.md
@@ -1,3 +1,13 @@
+### [0.3.16](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.15...@codecademy/markdown-overrides@0.3.16) (2020-11-30)
+
+### [0.3.15](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.14...@codecademy/markdown-overrides@0.3.15) (2020-11-30)
+
+### [0.3.14](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.13...@codecademy/markdown-overrides@0.3.14) (2020-11-19)
+
+### [0.3.13](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.12...@codecademy/markdown-overrides@0.3.13) (2020-11-19)
+
+### [0.3.12](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.11...@codecademy/markdown-overrides@0.3.12) (2020-11-19)
+
 ### [0.3.11](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.10...@codecademy/markdown-overrides@0.3.11) (2020-11-16)
 
 ### [0.3.10](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.3.9...@codecademy/markdown-overrides@0.3.10) (2020-11-13)